Consolidate MXFP4 GEMM dispatch and remove dead code
Move FP4 AITER GEMM handler from module/fp4_handler_gemm.py into
cpp_extensions/gemm.py alongside other GEMM dispatch paths. Remove
mxfp4_hip.cpp which is no longer needed after the nvte_quantize_v2
refactor.
